Refatorando servidor HTTP
